Initial changes to prepare files for the package/directory migration.
The build.gradle.kts files in identity, identity-flow, identity-mdoc, and
identity-sdjwt are all being merged, and that's easiest for the migration tool
if they're all identitical in the first place. This changes them so they're all
identical and will work after the migration.

The other changes are also preparation for the automatic migration.

In themselves, these changes break the build. However, after running the
migration script, these have been tested:

Tested by:
- ./gradlew check
- ./gradlew test
- ./gradlew connectedCheck
- Make Project in Android Studio
- Running testapp:
  - Adding documents
  - Testing mdoc sharing over NFC
  - Testing mdoc sharing using a QR code
  - TODO: Multi-device testing
- Running wallet:
  - Adding a Utopia PID
  - Adding a Utopia mDL
  - Verifying each using testapp, shared by QR code
  - Verifying each using the verifier web servlet

Signed-off-by: Kevin Deus <kdeus@google.com>